Delete Firmware directory
I have attempted to use the provided firmware for the SKR 3 (STM32H723VG) for the last 7 months with zero success. the Provided Marlin is absolute junk and will not compile as provided with any effort or settings adjustments. Even AI Copilot is unable to resolve this broken firmware and which leaves the product useless. 

I downloaded the newest version of Marlin which after some work does compile and leaves the unit is a infinite boot loop that continuously states "EEPROM Corrupt" which also can not be corrected. I have tested it on 2 separate SKR 3, and 7 difference PCs, with 40 different SD Cards, as well as attempted to Upload directly via USB. 

the Brand BIG TREE TECH, seems to be a China based knock off company that relies on other companies firmware and software releases to power their products. They have no official website and offer no actual customer service. leaving customers like myself to feel dupped into being ripped off on boards. 

 after trying for over a year to use their products on our CoreXY printers, we will be going back to the reliability and ease of use of MKS Robins, as they only require a simply .txt file to update and reconfigure and do not require the entire system to have a new OS loaded for every reconfiguration.